OP-3214 - moved get_asset_hierarchy_tree to tools folder
Shouldn't be in a global library, it is useful only for tools
